+++ packages/aws-client/src/types.ts
@@ -2,6 +2,7 @@ import * as AWS from 'aws-sdk';
 
 import { APIGatewayClient } from '@aws-sdk/client-api-gateway';
 import { CloudWatchLogs } from '@aws-sdk/client-cloudwatch-logs';
+import { CloudWatchEvents } from '@aws-sdk/client-cloudwatch-events';
 import { CloudFormation } from '@aws-sdk/client-cloudformation';
 import { DynamoDBStreamsClient } from '@aws-sdk/client-dynamodb-streams';
 import { DynamoDB, DynamoDBClient } from '@aws-sdk/client-dynamodb';
@@ -11,6 +12,6 @@ import { S3 } from '@aws-sdk/client-s3';
 import { SNS } from '@aws-sdk/client-sns';
 import { SQS } from '@aws-sdk/client-sqs';
 
-export type AWSClientTypes = APIGatewayClient | CloudFormation | CloudWatchLogs |
-DynamoDB | DynamoDBClient | DynamoDBStreamsClient | Kinesis | Lambda | S3 | SNS | SQS
-| AWS.Service | AWS.DynamoDB.DocumentClient;
+export type AWSClientTypes = APIGatewayClient | CloudFormation | CloudWatchLogs
+| CloudWatchEvents | DynamoDB | Kinesis | Lambda | S3 | SNS | SQS
+| DynamoDBClient | DynamoDBStreamsClient | AWS.Service | AWS.DynamoDB.DocumentClient;