+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- **CUMULUS-2888**
+  - Update CloudWatch Events code to AWS SDK v3
 - **CUMULUS-2889**
   - Updated Cloudwatch Logs code to aws sdk v3
 - **CUMULUS-2893**